Updated the render method
- rendering cannot be done more than once now, unless the view is not
visible.

Just as a side note, you should render a view once, and then update the
view many times over depending on what needs to be updated.  Render is
just a do-it once type of operation.  Perhaps in the future I can look
at the possibility of render being done many times as a sort of "redraw
everything and disregard everything else" type of operation.